+++ demo_nodes_cpp/src/parameters/even_parameters_node.cpp
@@ -14,6 +14,7 @@
 
 #include <vector>
 
+#include "rcl_interfaces/msg/set_parameters_result.hpp"
 #include "rclcpp/rclcpp.hpp"
 #include "rclcpp_components/register_node_macro.hpp"