+++ quality_of_service_demo/rclpy/quality_of_service_demo_py/qos_overrides_listener.py
@@ -12,6 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import sys
+
 import rclpy
 from rclpy.executors import ExternalShutdownException
 from rclpy.node import Node
@@ -54,9 +56,11 @@ def main(args=None):
     except (KeyboardInterrupt, ExternalShutdownException):
         pass
     finally:
-        rclpy.try_shutdown()
         node.destroy_node()
+        rclpy.try_shutdown()
+
+    return 0
 
 
 if __name__ == '__main__':
-    main()
+    sys.exit(main())